% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ac}
\alias{ac}
\title{Calculate the autocorrelation of a vector}
\usage{
ac(x, nLags = 50)
}
\arguments{
\item{x}{a vector storing parameter values}

\item{nLags}{the maximum number of lags}
}
\value{
A data.frame
}
\description{
Calculate the autocorrelation of a vector.
}
\examples{
res <- ac(1:100)
## List of 2
## $ Lag            : int [1:50] 1 2 3 4 5 6 7 8 9 10 ...
## $ Autocorrelation: num [1:50] 1 1 1 1 1 1 1 1 1 1 ...

res <- ac(rnorm(100))
str(res)
## List of 2
## $ Lag            : int [1:50] 1 2 3 4 5 6 7 8 9 10 ...
## $ Autocorrelation: num [1:50] 1 -0.0485 0.0265 -0.1496 0.0437 ...
}
