% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_pp}
\alias{assign_pp}
\title{Slot pp values into p.prior}
\usage{
assign_pp(pp, prior)
}
\arguments{
\item{pp}{pp.prior}

\item{prior}{p.prior}
}
\description{
Assign hyper-prior distributions to prior distributions
}
