% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autocorr}
\alias{autocorr}
\title{Autocorrelation Plot}
\usage{
autocorr(x, start = 1, end = NA, nLags = 50, pll = TRUE,
  subchain = FALSE)
}
\arguments{
\item{x}{posterior samples}

\item{start}{start from which iteration.}

\item{end}{end at which iteration}

\item{nLags}{the maximum number of lags.}

\item{pll}{a Boolean switch for plotting parameter values or posterior
log likelihoods}

\item{subchain}{a Boolean switch to plot a subset of chains.}
}
\description{
Plot the autocorrelation of posterior samples,
}
\examples{
## Model 1
## 27 elements with 20 levels
FR <- list(S = c("n","w","p"), cond=c("C","F", "H"), R=c("N", "W", "P"))
lev <- c("CnN","CwN", "CnW","CwW",
         "FnN","FwN","FpN", "FnW","FwW","FpW", "fa","FpP",
         "HnN","HwN","HpN", "HnW","HwW","HpW", "HpP",
         "FAKERATE")
map_mean_v <- ggdmc:::MakeEmptyMap(FR, lev)
map_mean_v[1:27] <- c(
  "CnN","CwN","FAKERATE", "FnN","FwN","FpN", "HnN","HwN","HpN",
  "CnW","CwW","FAKERATE", "FnW","FwW","FpW", "HnW","HwW","HpW",
  "FAKERATE","FAKERATE","FAKERATE", "fa","fa","FpP", "fa","fa","HpP")

model0 <- BuildModel(
  p.map     = list(A = "1", B = c("cond", "R"), t0 = "1", mean_v = c("MAPMV"),
                   sd_v = "1", st0 = "1", N = "cond"),
  match.map = list(M = list(n = "N", w = "W", p = "P"), MAPMV = map_mean_v),
  factors   = list(S = c("n","w","p"), cond = c("C","F", "H")),
  constants = c(N.C = 2, N.F = 3, N.H = 3, st0 = 0, B.C.P = Inf,
                mean_v.FAKERATE = 1, sd_v = 1),
  responses = c("N", "W", "P"),
  type      = "norm")

npar <- model0@npar

p.vector <- c(A = .3, B.C.N = 1.3,  B.F.N = 1.3,  B.H.N = 1.3,
              B.C.W = 1.3,  B.F.W = 1.4,  B.H.W = 1.5,
              B.F.P = 1.1,  B.H.P = 1.3,

              t0=.1,

              mean_v.CnN = 2.8,  mean_v.CwN = -0.3, mean_v.CnW=-1,
              mean_v.CwW = 2.9,  mean_v.FnN = 2.8,  mean_v.FwN=-.3,

              mean_v.FpN = -1.6, mean_v.FnW = -1,   mean_v.FwW = 2.9,
              mean_v.FpW = .5 ,  mean_v.fa = -2.4,  mean_v.FpP = 2.5,

              mean_v.HnN = 2.8, mean_v.HwN = -.5,   mean_v.HpN = -.6,
              mean_v.HnW = -.7, mean_v.HwW = 3.0,   mean_v.HpW = 1.6,
              mean_v.HpP = 2.3)

acc_tab0 <- TableParameters(p.vector, 1, model0, FALSE)
acc_tab1 <- TableParameters(p.vector, "w.C.N", model0, FALSE)
acc_tab2 <- TableParameters(p.vector, "w.F.P", model0, FALSE)
print(acc_tab0); print(acc_tab1); print(acc_tab2)

\dontrun{
dat0 <- simulate(model0, nsim=50, ps=p.vector)
dmi0 <- BuildDMI(dat0, model0)
}
p1 <- rep(1, npar)
names(p1) <- model0@pnames

p.prior0 <- BuildPrior(
  dists = c(rep("tnorm", 9), "beta", rep("tnorm", 19)),
  p1    = p1,
  p2    = c(rep(2, 9), 1, rep(2, 19)),
  lower = c(rep(0, 10),  rep(NA, 19)),
  upper = c(rep(NA, 9), 1, rep(NA, 19)))

# plot(p.prior0, ps = p.vector)
## Sampling
## 18.4 & 36.17 s
\dontrun{
fit0 <- StartNewsamples(dmi0, p.prior0, block = FALSE, thin=4)
fit0_correct  <- run(fit0, thin=4, block = FALSE)

hat <- gelman(fit0_correct, verbose=TRUE);
p0 <- autocorr(fit0_correct, subchain=1:3, pll=TRUE)
}
}
