% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc_model.R
\name{check_dmi}
\alias{check_dmi}
\alias{BindDataModel}
\title{Bind Data and Models}
\usage{
check_dmi(data, model)

BindDataModel(data, model, n.pda = 16384, bw = 0.01, gpuid = 0,
  debug = FALSE)
}
\arguments{
\item{data}{a data frame stored choice-RT data}

\item{model}{a DMC model}

\item{bw}{bandwidth}

\item{gpuid}{indicate using which GPU card. Default is 0.}

\item{debug}{if print debugging information}

\item{npda}{number of model simulations}
}
\description{
Binding a data frame and a DMC model description. This function also check
if they are compatible and adding a \code{cell.index} and many other
attributes to the data frame in order to speed likelihood computation.
}
\examples{
model <- BuildModel(
   p.map     = list(a="1",v="1",z="1",d="1",sz="1",sv="1", t0="1",st0="1"),
   constants = c(st0=0,d=0),
   match.map = list(M=list(s1="r1",s2="r2")),
   factors   = list(S=c("s1","s2")),
   responses = c("r1","r2"),
   type      = "rd")

p.vector <- c(a=1,v=1, z=0.5, sz=0.25, sv=0.2,t0=.15)
dat <- simulate(model, p.vector, 128)
dmi <- BindDataModel(model, dat)

}
