% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{BuildDMI}
\alias{BuildDMI}
\title{Bind data and models}
\usage{
BuildDMI(x, model)
}
\arguments{
\item{x}{data formatted as a data frame}

\item{model}{a model object}
}
\value{
a data model instance
}
\description{
Binding a data set with an object of data-model instance. The function
checks whether the data and the model are compatible and adds attributes
to a data model instance.
}
