% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{check_pvec}
\alias{check_pvec}
\title{Check if parameter vector compatible with model object?}
\usage{
check_pvec(x, model)
}
\arguments{
\item{x}{parameter vector}

\item{model}{a model object}
}
\description{
Check if the user supplies a parameter vector, compatiable with
the model object.
}
