% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checkddm2}
\alias{checkddm2}
\title{Check parameter vector of DDM}
\usage{
checkddm2(pVec)
}
\arguments{
\item{pVec}{a numeric vector storing the DDM parameters}
}
\description{
Check if all DDM parameters are within reasonable ragnes
}
