% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckDMI}
\alias{CheckDMI}
\title{Check Data Model Instance}
\usage{
CheckDMI(x = NULL, prior = NULL, theta1 = NULL, nchain = NULL)
}
\arguments{
\item{x}{a data-model instance}

\item{prior}{a parameter prior list}

\item{theta1}{a user-supplied theta cube}

\item{nchain}{number of MCMC chains}
}
\description{
Return a model object extracted either from a data model instance or
an object storing posterior samples. The function checks also
\enumerate{
\item When x stores DMI of one participant, if DMI is a \code{data.frame},
\item When x is an object of posterior samples, if x is a list of many subjects,
\item whether model is successfully created,
\item whether prior is suppled or we can extract it from 'samples'
}
}
