% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckRJ}
\alias{CheckRJ}
\title{Check rejection rate}
\usage{
CheckRJ(object, verbose = TRUE)
}
\arguments{
\item{object}{posterior samples}

\item{verbose}{print more information}
}
\description{
Check rejection rate for posterior samples
}
