% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConvertChains}
\alias{ConvertChains}
\title{Prepare posterior samples for plotting functions version 1}
\usage{
ConvertChains(x, start = 1, end = NA, pll = TRUE)
}
\arguments{
\item{x}{posterior samples}

\item{start}{which iteration to start}

\item{end}{end at which iteration}

\item{pll}{a Boolean switch to make posterior log likelihood}
}
\description{
Convert MCMC chains to a data frame for plotting functions
}
