% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConvertChains2}
\alias{ConvertChains2}
\title{Prepare posterior samples for plotting version 2}
\usage{
ConvertChains2(x, pll)
}
\arguments{
\item{x}{posterior samples}

\item{pll}{a Boolean switch to make posterior log likelihood}
}
\description{
Convert MCMC chains to a data frame for plotting
}
