% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{dbeta_lu}
\alias{dbeta_lu}
\alias{rbeta_lu}
\title{A slightly modified dbeta and rbeta functions}
\usage{
dbeta_lu(x, shape1, shape2, lower, upper, log = FALSE)

rbeta_lu(n, shape1, shape2, lower, upper)
}
\arguments{
\item{x}{posterior samples}

\item{shape1}{shape1 parameter}

\item{shape2}{shape1 parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{log}{log density?}

\item{n}{number of simulations}
}
\description{
A slightly modified dbeta and rbeta functions
}
