% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{dcauchy_l}
\alias{dcauchy_l}
\title{A modified dcauchy functions}
\usage{
dcauchy_l(x, p1, p2, lower, upper, lg = FALSE)
}
\arguments{
\item{x}{quantile}

\item{p1}{location parameter}

\item{p2}{scale parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{lg}{log density?}
}
\description{
A modified dcauchy functions
}
