% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{dconstant}
\alias{dconstant}
\alias{rconstant}
\title{A pseudo constant function to get constant densities}
\usage{
dconstant(x, constant, log = FALSE)

rconstant(n, constant)
}
\arguments{
\item{x}{posterior samples}

\item{constant}{constant value}

\item{log}{log density?}

\item{n}{number of simulations}
}
\description{
Used with constant prior
}
