% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{deviance.model}
\alias{deviance.model}
\title{Bayesian measures of model complexity and fit}
\usage{
\method{deviance}{model}(object, ...)
}
\arguments{
\item{object}{posterior samples}

\item{...}{other plotting arguments passing through dot dot dot.}
}
\description{
Calculate deviance for one fitted model objects for which a
log-likelihood value can be obtained, according to the formula
-2*log-likelihood ... unfinished ...
}
