% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{dgamma_l}
\alias{dgamma_l}
\alias{rgamma_l}
\title{A slightly modified dgamma and rgamma functions}
\usage{
dgamma_l(x, shape, scale, lower, upper, log = FALSE)

rgamma_l(n, shape, scale, lower, upper)
}
\arguments{
\item{x}{posterior samples}

\item{shape}{shape parameter}

\item{scale}{scale parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{log}{log density?}

\item{n}{number of simulations}
}
\description{
A slightly modified dgamma and rgamma functions
}
