% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{dlnorm_l}
\alias{dlnorm_l}
\alias{rlnorm_l}
\title{A slightly modified dlnorm and rlnorm functions}
\usage{
dlnorm_l(x, meanlog, sdlog, lower, upper, log = FALSE)

rlnorm_l(n, meanlog, sdlog, lower, upper)
}
\arguments{
\item{x}{posterior samples}

\item{meanlog}{meanlog parameter}

\item{sdlog}{sdlog parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{log}{log density?}

\item{n}{number of simulations}
}
\description{
A slightly modified dlnorm and rlnorm functions
}
