% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{.calculate_between_chain_stats}
\alias{.calculate_between_chain_stats}
\title{Calculate between-chain statistics}
\usage{
.calculate_between_chain_stats(xbar, niter)
}
\arguments{
\item{xbar}{Matrix of chain means (parameters × chains)}

\item{niter}{Number of iterations used}
}
\value{
Between-chain covariance matrix
}
\description{
Internal function to compute between-chain variances
}
\keyword{internal}
