% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{.gelman_diag}
\alias{.gelman_diag}
\title{Brook-Gelman diagnosis}
\usage{
.gelman_diag(x, start = 1, end = NA, conf = 0.95, digits = 2)
}
\arguments{
\item{x}{A \code{posterior} object or list of posterior samples}

\item{start}{First iteration to use}

\item{end}{Last iteration to use (NA for all)}

\item{conf}{Confidence level for upper CI (default = 0.95)}

\item{digits}{Number of digits to print (default = 2)}
}
\value{
List with psrf and mpsrf
}
\description{
Internal function for hierarchical models
}
\keyword{internal}
