% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{.get_multivariate_psrf}
\alias{.get_multivariate_psrf}
\title{Get multivariate PSRF}
\usage{
.get_multivariate_psrf(npar, niter, W, B, multivariate)
}
\arguments{
\item{npar}{Number of parameters}

\item{W}{Within-chain covariance matrix}

\item{B}{Between-chain covariance matrix}
}
\value{
Multivariate PSRF or NULL
}
\description{
Internal function for multivariate R-hat
}
\keyword{internal}
