% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dprior}
\alias{dprior}
\alias{sumlogpriorNV}
\title{Probability densities of prior distributions}
\usage{
dprior(pvec, prior)

sumlogpriorNV(pvec, prior)
}
\arguments{
\item{pvec}{parameter vector}

\item{prior}{prior distributions}
}
\value{
a vector
}
\description{
\code{sumlogpriorNV} calculate sum log-likelihood. \code{rprior} generates
random observation from prior distributions
}
