% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc_analysis.R
\name{Dstats.ddm}
\alias{Dstats.ddm}
\title{Calculate Dstats of DDM Density}
\usage{
Dstats.ddm(samples, save = FALSE, fast = TRUE)
}
\arguments{
\item{samples}{a DMC sample/object}

\item{save}{a save switch}

\item{fast}{choose different calculation routine}
}
\description{
Calculates and returns list with mean (meanD), variance (varD) and min
(minD) of Deviance, deviance of mean theta (Dmean) and (optionally) matrix
of deviances for each theta.
}
\details{
This function is still under tweaking. Please use DMC's \code{Dstat.dmc},
instead.
}
\keyword{Dstats}

