% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{effectiveSize}
\alias{effectiveSize}
\alias{effectiveSize,hyper-method}
\alias{effectiveSize,list-method}
\alias{effectiveSize,posterior-method}
\title{Effective Sample Size}
\usage{
effectiveSize(x, ...)

\S4method{effectiveSize}{hyper}(x, hyper = TRUE, start = 1, end = NA,
  subchain = NA, digits = 2, verbose = FALSE)

\S4method{effectiveSize}{list}(x, start = 1, end = NA, subchain = NA,
  digits = 2, verbose = FALSE)

\S4method{effectiveSize}{posterior}(x, start = 1, end = NA,
  subchain = NA, digits = 2, verbose = FALSE)
}
\arguments{
\item{x}{posterior samples}

\item{...}{other additional arguments}

\item{hyper}{a Boolean switch to calculate phi}

\item{start}{start from iteration}

\item{end}{end at which iteraton}

\item{subchain}{calculate a subset of chains. This must be an integer vector}

\item{digits}{printing how many digits}

\item{verbose}{printing more information}
}
\description{
Posterior sample size adjusted for autocorrelation. The function is based on
the effectiveSize function in \code{coda} package.
}
\details{
hyper argument does not work for list class (i.e., posterior samples from
a fixed-effect model fit).
}
\examples{
#################################40
## effectiveSize example
#################################40
\dontrun{
cat("Class:", class(fit), "\\n")
es1 <- effectiveSize(fit, hyper=TRUE, verbose=FALSE)
es1 <- effectiveSize(fit, hyper=TRUE, verbose=TRUE)

es1 <- effectiveSize(fit, hyper=TRUE, verbose=FALSE, subchain=7:9)
es1 <- effectiveSize(fit, hyper=TRUE, verbose=TRUE, subchain=7:9)

es1 <- effectiveSize(fit, hyper=FALSE, verbose=FALSE)
es1 <- effectiveSize(fit, hyper=FALSE, verbose=TRUE)

es1 <- effectiveSize(fit, hyper=FALSE, verbose=FALSE, subchain=4:6)
es1 <- effectiveSize(fit, hyper=FALSE, verbose=TRUE, subchain=4:6)

cat("Starting a new fixed-effect model fit: \\n")
fit0 <- StartNewsamples(dmi, p.prior, ncore=4)
fit  <- run(fit0, ncore=4)

cat("Class:", class(fit), "\\n")
es1 <- effectiveSize(fit, verbose=FALSE)
es1 <- effectiveSize(fit, verbose=TRUE)
es1 <- effectiveSize(fit, verbose=FALSE, subchain=4:6)
es1 <- effectiveSize(fit, verbose=TRUE, subchain=4:6)

}

}
\references{
\enumerate{
Plummer, M. Best, N., Cowles, K., Vines, K., Sarkar, D., Bates, D., Almond, R., & Magnusson, A. (2019). R package 'coda' \url{https://cran.r-project.org/web/packages/coda/}
}
}
