% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{gelman}
\alias{gelman}
\alias{gelman,posterior-method}
\alias{gelman,list-method}
\alias{gelman,hyper-method}
\title{Potential scale reduction factor}
\usage{
gelman(x, ...)

\S4method{gelman}{posterior}(x, start = 1, end = NA, conf = 0.95,
  multivariate = TRUE, subchain = NA, digits = 2, verbose = FALSE)

\S4method{gelman}{list}(x, start = 1, end = NA, conf = 0.95,
  multivariate = TRUE, subchain = NA, digits = 2, verbose = FALSE)

\S4method{gelman}{hyper}(x, hyper = TRUE, start = 1, end = NA,
  conf = 0.95, multivariate = TRUE, subchain = NA, digits = 2,
  verbose = FALSE)
}
\arguments{
\item{x}{posterior samples}

\item{...}{other additional arguments}

\item{start}{start iteration}

\item{end}{end iteration}

\item{conf}{confident inteval}

\item{multivariate}{multivariate Boolean switch}

\item{subchain}{whether only calculate a subset of chains}

\item{digits}{print out how many digits}

\item{verbose}{print more information}

\item{hyper}{a Boolean switch, indicating posterior samples are from
hierarchical modeling}
}
\description{
\code{gelman} function calls the function, \code{gelman.diag} in the
\pkg{coda} package to calculates PSRF.
}
\examples{
\dontrun{
rhat1 <- gelman(hsam);
rhat2 <- gelman(hsam, end = 51);
rhat3 <- gelman(hsam, conf = .90);
rhat7 <- gelman(hsam, subchain = TRUE);
rhat8 <- gelman(hsam, subchain = 1:4);
rhat9 <- gelman(hsam, subchain = 5:7, digits = 1, verbose = TRUE);
}
}
