% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{gelman}
\alias{gelman}
\alias{gelman-methods}
\alias{gelman,posterior-method}
\title{Brook-Gelman Diagnostic (R-hat)}
\usage{
gelman(x, start = 1, end = NA, conf = 0.95)

\S4method{gelman}{posterior}(x, start = 1, end = NA, conf = 0.95)
}
\arguments{
\item{x}{A \code{posterior} object or list of posterior samples}

\item{start}{First iteration to use}

\item{end}{Last iteration to use (NA for all)}

\item{conf}{Confidence level for upper CI (default = 0.95)}
}
\value{
A list containing psrf and mpsrf (if requested)
}
\description{
Calculates potential scale reduction factors for MCMC chains.
}
\examples{
\dontrun{
fits <- fits1
phi <- RebuildHyper(fits)
thetas <- RebuildPosteriors(fits)
hat <- gelman(phi)
hat <- gelman(fits[[1]]$phi)
hat <- gelman(fits[[2]]$phi)
hat <- gelman(fits[[3]]$phi)
}
}
