% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc_hierarchical.R
\name{GetNsim}
\alias{GetNsim}
\title{Return n-cell matrix}
\usage{
GetNsim(model, n, ns)
}
\arguments{
\item{model}{a model object}

\item{n}{number of trials/responses.}

\item{ns}{number of subjects.}
}
\description{
Contructs a matrix indicating how many responses to simulate in
each design cell. It also check whether the format of \code{n} and \code{ns}
conform.
}
\details{
\code{n} can be (1) one real number for a balanced design, (2) a matrix for
an unbalanced design, where rows are subjects and columns are design cells.
If the matrix is a row vector, then all subjects have the same \code{n} in
each cell. If it is a column vector, then all cells have the same \code{n}.
Otherwise each entry specifies the \code{n} for a particular design
subject x design cell combination. See examples below.
}
\examples{
model <- ggdmc::BuildModel(
p.map       = list(A="1",B="R",t0="1",mean_v="M",sd_v="M",st0="1"),
  match.map = list(M=list(s1=1,s2=2)),
  constants = c(sd_v.false=1,st0=0),
  factors   = list(S=c("s1","s2")),
  responses = c("r1","r2"),
  type      = "norm")

## Example 1
ggdmc::GetNsim(model, ns = 2, n = 1)
#      [,1] [,2]
# [1,]    1    1
# [2,]    1    1

## Example 2
n <- matrix(c(1:2),ncol=1)
#      [,1]
# [1,]    1  ## subject 1 has 1 response for each cell
# [2,]    2  ## subject 2 has 2 responses for each cell

ggdmc::GetNsim(model, ns = 2, n = n)
#      [,1] [,2]
# [1,]    1    1
# [2,]    2    2

## Example 3
n <- matrix(c(1:2), nrow=1)
#      [,1] [,2]
# [1,]    1    2
ggdmc::GetNsim(model, ns = 2, n = n)
#     [,1] [,2]
# [1,]   1    2 ## subject 1 has 1 response for cell 1 and 2 responses for cell 2
# [2,]   1    2 ## subject 2 has 1 response for cell 1 and 2 responses for cell 2

## Example 4
n <- matrix(c(1:4), nrow=2)
#      [,1] [,2]
# [1,]    1    3
# [2,]    2    4
ggdmc::GetNsim(model, ns = 2, n = n)
#      [,1] [,2]
# [1,]    1    3 ## subject 1 has 1 response for cell 1 and 3 responses for cell 2
# [2,]    2    4 ## subject 2 has 2 responses for cell 1 and 4 responses for cell 2
}
