% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetPNames}
\alias{GetPNames}
\title{Extract parameter names from a model object}
\usage{
GetPNames(x)
}
\arguments{
\item{x}{a model object}
}
\description{
GetPNames will be deprecated. Please extract pnames directly via S4 slot 'model@pnames'
}
