% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetTheta0}
\alias{GetTheta0}
\title{Extract Start Posterior Sample}
\usage{
GetTheta0(samples)
}
\arguments{
\item{samples}{posterior samples}
}
\value{
a nsubject x npar x nchain array
}
\description{
Extract the theta's of the first MCMC iteration across chains and
participants. Note that the ps array in DMC is a nchain x nsubject x
nparameter array. Armadillo operates on slice (the third dimension), so
chain dimension has to be on slice.
}
