% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hsumloglike}
\alias{hsumloglike}
\title{Add log-likelihoods across subjects at the hyper level}
\usage{
hsumloglike(ps, pp, prior)
}
\arguments{
\item{ps}{a nsubject x npar matrix}

\item{pp}{a temporary pp.prior values extracted from phi. The temporary
pp.prior has no attached parameter names. phi is two-element list. First
element is a location array of nchain x npar x nmc; second element is a
scale array of nchain x npar x nmc.}

\item{prior}{prior distributions at data level}
}
\description{
An external sum log likelihoods for hyper parameters.
}
