% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{hyper-class}
\alias{hyper-class}
\title{An S4 class to represent an object storing posterior samples at the
participant and hyper level}
\description{
An S4 class to represent an object storing posterior samples at the
participant and hyper level
}
\section{Slots}{

\describe{
\item{\code{phi_loc}}{posterior samples for the location parameters}

\item{\code{phi_sca}}{posterior samples for the scale parameters}

\item{\code{summed_log_prior}}{summed log prior likelihoods for phi.}

\item{\code{log_likelihoods}}{log likelihoods for phi}

\item{\code{prior_loc}}{a S4 prior object for the location parameters}

\item{\code{prior_sca}}{a S4 prior object for the scale parameters}

\item{\code{start}}{the index of starting sample}

\item{\code{npar}}{number of parameters}

\item{\code{pnames}}{parameter names}

\item{\code{nmc}}{number of Monte Carlo samples}

\item{\code{thin}}{thinning length}

\item{\code{nchain}}{number of Markov chains}

\item{\code{individuals}}{a list storing posterior samples for each individual participant}

\item{\code{snames}}{names of individual participants}
}}

\seealso{
\code{\link{posterior-class}}
}
