% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialise_hyper}
\alias{initialise_hyper}
\title{Set up a DMC Sample for Multiple Participants}
\usage{
initialise_hyper(nmc, pList = NULL, data = NULL, samples = NULL,
  theta1 = NULL, startPList = NULL, phi1 = NULL, ppList = NULL,
  hStartPList = NULL, setting = NULL)
}
\arguments{
\item{nmc}{number of Markov Chain Monte Carlo iteration}

\item{pList}{prior distribution setting. This is R's p.prior}

\item{data}{a model data instance created by \code{data.model.dmc}}

\item{samples}{a DMC posterior sample}

\item{theta1}{A user supplied initial theta cube}

\item{startPList}{A user supplied (different) prior distribution setting.
This is R's start.prior.}

\item{phi1}{A user supplied initial phi cube}

\item{ppList}{prior distribution setting, hyper level}

\item{hStartPList}{A user supplied (different) hyper-prior distribution
setting}

\item{setting}{a list container to store all DMC setting}
}
\description{
This is C++ function to initialse a DMC sample with multiple participants.
The user usually should not directly use it.
}

