% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{iseffective}
\alias{iseffective}
\title{Four pMCMC checking tools for automatic sampling}
\usage{
iseffective(x, minN, nfun, verbose = FALSE)
}
\arguments{
\item{x}{posterior samples}

\item{minN}{minimal effective sample sizes}

\item{nfun}{mean or median function}

\item{verbose}{print more information}
}
\description{
The function tests whether MCMC chains have drawn enough samples.
}
