% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{isflat}
\alias{isflat}
\title{Model checking functions}
\usage{
isflat(x, p1 = 1/3, p2 = 1/3, cut_location = 0.25, cut_scale = Inf,
  verbose = FALSE)
}
\arguments{
\item{x}{posterior samples}

\item{p1}{the range of the head of MCMC chains}

\item{p2}{the range of the tail of the MCMC chains}

\item{cut_location}{how far away a location chains been considered as stuck}

\item{cut_scale}{how far away a scale chains been considered as stuck}

\item{verbose}{print more information}
}
\description{
The function tests whether Markov chains converge prematurelly:
}
