% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ismanymodels}
\alias{ismanymodels}
\title{Test whether `model` object has many models}
\usage{
ismanymodels(model, ns = NA)
}
\arguments{
\item{model}{a model object}

\item{ns}{number of subjects.}
}
\description{
Test whether `model` object has many models
}
