% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ismixed}
\alias{ismixed}
\title{Model checking functions}
\usage{
ismixed(x, cut = 1.01, split = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{posterior samples}

\item{cut}{psrf criterion for well mixed}

\item{split}{whether to split MCMC chains. This is an argument passing to
\code{gelman} function}

\item{verbose}{print more information}
}
\description{
The function tests whether Markov chains are mixed well.
}
\seealso{
\code{\link{gelman}})
}
