% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{isstuck}
\alias{isstuck}
\alias{CheckConverged}
\title{Four model checking tools for automatic sampling}
\usage{
isstuck(x, hyper = FALSE, cut = 10, start = 1, end = NA,
  verbose = FALSE)

CheckConverged(x)
}
\arguments{
\item{x}{posterior samples}

\item{hyper}{a Boolean switch, extracting hyper attribute.}

\item{cut}{the criteria for suggesting abnormal chains found}

\item{start}{start iteration}

\item{end}{end iteration}

\item{verbose}{print more information}
}
\description{
The function tests whether MCMC chains encounter a parameter region
difficult to search (ie get stuck):
}
