% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmc_eam.R
\name{likelihood.ddm}
\alias{likelihood.ddm}
\alias{likelihood.lba_B}
\title{Calculate Log-Likelihood}
\usage{
likelihood.ddm(p.vector, data, min.like = 1e-10, ...)

likelihood.lba_B(p.vector, data, min.like = 1e-10, ...)
}
\arguments{
\item{p.vector}{a parameter vector}

\item{data}{a model data instance}

\item{min.like}{minimal likelihood, 1e-10}

\item{...}{other parameters, such as \emph{precision} for \code{ddm},
\emph{ok.types} for \code{lba_B}.}
}
\value{
a vector of likelihoods for each RT in data (in same order)
}
\description{
\code{likelihood.ddm} is a legacy function for computing diffusion
model density. If the speed of computation is a concern, the user may want to
use \code{ddmc}, which implemnts a streamlined C++ diffusion model density.
}
\keyword{likelihood}

