% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{logLik}
\alias{logLik}
\alias{logLik,posterior-method}
\alias{logLik,list-method}
\alias{logLik,hyper-method}
\title{Extract Posterior Log-Likelihood}
\usage{
logLik(object, ...)

\S4method{logLik}{posterior}(object, start = 1, end = NA)

\S4method{logLik}{list}(object, start = 1, end = NA)

\S4method{logLik}{hyper}(object, start = 1, end = NA)
}
\arguments{
\item{object}{posterior samples}

\item{...}{other arguments passing through dot dot dot.}

\item{start}{start from which iteration.}

\item{end}{end at which iteration. For example, set
\code{start = 101} and \code{end = 1000}, instructs the function to
calculate from 101 to 1000 iteration.}

\item{hyper}{whether to summarise hyper parameters}
}
\description{
This function is to extract posterior log-likelihood in a "model" object.
}
