% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{logLik.model}
\alias{logLik.model}
\title{Extract Posterior Log-Likelihood}
\usage{
\method{logLik}{model}(object, hyper = FALSE, start = 1, end = NA,
  ...)
}
\arguments{
\item{object}{posterior samples}

\item{hyper}{whether to summarise hyper parameters}

\item{start}{start from which iteration.}

\item{end}{end at which iteration. For example, set
\code{start = 101} and \code{end = 1000}, instructs the function to
calculate from 101 to 1000 iteration.}

\item{...}{other arguments passing through dot dot dot.}
}
\description{
This function is to extract posterior log-likelihood in a "model" object.
}
\examples{

}
