% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc_model.R
\name{make.level.array}
\alias{make.level.array}
\title{Create an array of all factorial combinations of factor levels}
\usage{
make.level.array(factors = NA)
}
\arguments{
\item{factors}{a list storing one or multiple factor vector. Each vector
can have different numbers of level.  For example,
\code{f <- list(E = c("nc", "wc"), S = c("n", "w", "pn", "pw"))}}
}
\value{
a table showing the combinations of factor levels.
}
\description{
Take a list storing one or multiple string vectors and glues the strings in
these vectors with a dot symbol.
}
\examples{
## Example 1
factors <- list(S = c("s1", "s2"))
ggdmc:::make.level.array(factors)
##   S
##  s1   s2
## "s1" "s2"

factors <- list(S = c("left", "right"))
ggdmc:::make.level.array(factors)
##      S
##   left   right
##  "left" "right"

## Example 2
factors <- list(A = c("00", "45", "90", "135", "180"),
                S = c("mirror", "normal"))
ggdmc:::make.level.array(factors)
##       S
## -------------------------------
##   A        mirror       normal
## -------------------------------
##  00   "00.mirror"  "00.normal"
##  45   "45.mirror"  "45.normal"
##  90   "90.mirror"  "90.normal"
## 135  "135.mirror" "135.normal"
## 180  "180.mirror" "180.normal"

## Example 3
factors <- list(E = c("nc", "wc"),
                S = c("n", "w", "pn", "pw"))
ggdmc:::make.level.array(factors)

##         S
## ---------------------------------
## E       n      w      pn      pw
## ---------------------------------
## nc "nc.n" "nc.w" "nc.pn" "nc.pw"
## wc "wc.n" "wc.w" "wc.pn" "wc.pw"

}
