% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MakeForce}
\alias{MakeForce}
\title{Create a vector for re-calculation}
\usage{
MakeForce(x, nth)
}
\arguments{
\item{x}{posterior samples}

\item{nth}{every nth step to recalculate}
}
\description{
The is a PDA function. It creates an index vector, indicating which
iteration to recalculate likelihoods.
}
\examples{
\dontrun{
  MakeForce(hsam[[1]], 3)
}

}
