% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mcmc_list.model}
\alias{mcmc_list.model}
\title{Create a MCMC list}
\usage{
mcmc_list.model(x, start = 1, end = NA, pll = TRUE)
}
\arguments{
\item{x}{posterior samples}

\item{start}{start from which iteration}

\item{end}{end at which iteration}

\item{pll}{a Boolean switch for calculating posterior log-likelihood}
}
\description{
Create a MCMC list
}
