% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n1PDF_cnorm}
\alias{n1PDF_cnorm}
\title{Likelihood function for correlated accumulator model}
\usage{
n1PDF_cnorm(x, A, b, t0, mean_v, sd_v, st0, corr_v, n, h, debug = FALSE)
}
\arguments{
\item{x}{response time vector}

\item{A}{start point variability}

\item{b}{threshold}

\item{t0}{non-decision time}

\item{mean_v}{mean drift rate vector}

\item{sd_v}{standard deviation of the drift rates}

\item{st0}{non-decision time variability}

\item{corr_v}{correlation among accumulators}

\item{n}{number of simulations}

\item{h}{kernel bandwidth}

\item{debug}{a Boolean switch for debugging}
}
\description{
Calculate first node probability densities of the correlated accumulator
model
}
