% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n1PDF_gpu}
\alias{n1PDF_gpu}
\title{A Rcpp connection to the GPU-based LBA n1PDF in ppda package}
\usage{
n1PDF_gpu(x, A, b, mean_v, sd_v, t0, n, nthread, gpuid, bw, debug)
}
\arguments{
\item{x}{response time vector}

\item{A}{start point variability}

\item{b}{threshold}

\item{mean_v}{mean drift rate vector}

\item{sd_v}{standard deviation of the drift rates}

\item{t0}{nondecision time}

\item{n}{number of simulations}

\item{nthread}{thread number to launch in GPU}

\item{gpuid}{which gpu card to use, starting from 0.}

\item{bw}{kernel bandwidth}

\item{debug}{debugging?}
}
\description{
This is an Rcpp function calling a \code{n1PDF_gpu} function defined
in the global environment, which connects to an interanl ppda
simulated-based LBA density function.
}
