% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmc_plotting.R
\name{pairs.dmc}
\alias{pairs.dmc}
\title{Create a Plot Matrix of Posterior Simulations}
\usage{
\method{pairs}{dmc}(x, start = 1, ...)
}
\arguments{
\item{x}{a DMC sample/object}

\item{start}{plot from which iteration}

\item{...}{other arguments}
}
\description{
This is a wrapper function to plot a DMC object, using \pkg{ggmcmc}
\code{ggs_pairs}
}

