% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pairs.model}
\alias{pairs.model}
\title{Correlation Matrix}
\usage{
\method{pairs}{model}(x, start = 1, end = NA, ...)
}
\arguments{
\item{x}{posterior samples}

\item{start}{start from which iteration.}

\item{end}{end at which iteration.}

\item{...}{other arguments}
}
\description{
Plot the correlation matrix of posterior samples,
}
