% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc_analysis.R
\name{phi.as.mcmc.list}
\alias{phi.as.mcmc.list}
\title{Convert Phi to a Theta Vector}
\usage{
phi.as.mcmc.list(hyper, start = 1, end = NA)
}
\arguments{
\item{hyper}{a hyper attribute extracted from a hierarchical sample}

\item{start}{start iteration}

\item{end}{end iteration}
}
\description{
\code{phi.as.mcmc.list} extract phi from a sample's hyper attribute and
convert it to theta vector to a mcmc.list \pkg{coda} object
}

