% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc_analysis.R
\name{pick.stuck.dmc}
\alias{pick.stuck.dmc}
\title{Find Stuck Chains}
\usage{
pick.stuck.dmc(samples, hyper = FALSE, cut = 10, start = 1, end = NA,
  verbose = FALSE, digits = 2)
}
\arguments{
\item{samples}{a DMC object/sample}

\item{hyper}{whether to work on hyper parameters, namely phi}

\item{cut}{cutoff point. Default 10}

\item{start}{starting from which iteration. Default 1}

\item{end}{calculating until which iteration. Default is the last one,
extracted from the input DMC sample.}

\item{verbose}{whether print out debugging information}

\item{digits}{print how many digits. Default 2}
}
\value{
index of stuck chains (deviaton of ll from median < -cut)
}
\description{
This function picks out the chains that get stuck.
}
\keyword{pick.stuck}

