% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmc_plotting.R
\name{plot.pp.ggdmc}
\alias{plot.pp.ggdmc}
\title{Posterior Predictive Plot}
\usage{
\method{plot}{pp.ggdmc}(x, style = "pdf", gpvar1 = NULL, mode = "mode",
  size = 18, legPos = c(0.85, 0.85), legKeyS = unit(2, "lines"),
  legKeyW = unit(1, "cm"), xlim = c(0, Inf), ...)
}
\arguments{
\item{x}{post predictive object created usually by
\code{post.predict.ggdmc}.}

\item{style}{to plot probability density, \code{pdf}, cumulative
density, \code{cdf}, or \code{both}}

\item{gpvar1}{add one more experimental factor for faceting}

\item{mode}{an argument}

\item{size}{font size}

\item{legPos}{where to anchor legend, in percentage of the x and y axes}

\item{legKeyS}{legend key size}

\item{legKeyW}{legend key width}

\item{xlim}{x axis range}

\item{...}{other arguments}
}
\description{
This is posterior predictive plot ggplot2 version
}

