% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmc_utils.R
\name{post.predict.ggdmc}
\alias{post.predict.ggdmc}
\title{Simulate Post-predictive Sample}
\usage{
post.predict.ggdmc(samples, n.post = 100, probs = c(1:99)/100,
  bw = "nrd0", report = 10, save.dat = FALSE)
}
\arguments{
\item{samples}{a DMC sample/object}

\item{n.post}{how many data point to simulate}

\item{probs}{the quantile probability. Default c(1:99)/100}

\item{bw}{bin width method. Default "nrd0", passing to density function}

\item{report}{report progress}

\item{save.dat}{whether to save simulation data to further modify data
frame for plotting}
}
\description{
Make a list of probability density, quantiles and resposnes based on a
posterior DMC sample. This function is to produce a data frame that can be
used by ggplot2 functions.
}

