% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{posterior-class}
\alias{posterior-class}
\title{An S4 class to represent an object storing posterior samples.}
\description{
An S4 class to represent an object storing posterior samples.
}
\section{Slots}{

\describe{
\item{\code{theta}}{posterior samples.}

\item{\code{summed_log_prior}}{summed log prior likelihoods.}

\item{\code{log_likelihoods}}{logged likelihoods}

\item{\code{start}}{the index of starting sample}

\item{\code{npar}}{number of free parameters}

\item{\code{pnames}}{free parameter names}

\item{\code{nmc}}{number of Monte Carlo samples}

\item{\code{thin}}{thinning length}

\item{\code{nchain}}{number of Markov chains facilitating the DE-MCMC sampling}
}}

