% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{predict_one}
\alias{predict_one}
\title{Post-predictive Simulation}
\usage{
predict_one(object, npost = 100, rand = TRUE, factors = NA,
  xlim = NA, seed = NULL)
}
\arguments{
\item{object}{a model object of a subject.}

\item{npost}{number of posterior predictive replications.}

\item{factors}{experimental factors}

\item{xlim}{trimming outlier, e.g., xlim = c(0, 5).}

\item{seed}{an integer specifying if and how the random number generator
should be initialized.}

\item{nsub}{whether randomly select a npost estimated parameter values or
select the first npost estimates}
}
\value{
a data frame
}
\description{
Simulate post-predictive data.
}
