% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{prior-class}
\alias{prior-class}
\title{An S4 class to represent an object storing prior distributions}
\description{
An S4 class to represent an object storing prior distributions
}
\section{Slots}{

\describe{
\item{\code{npar}}{the number of parameters}

\item{\code{pnames}}{the names of parameters}

\item{\code{priors}}{a list storing the location parameter, scale parameter, upper
bound, lower bound, log indicator (0=FALSE, 1=TRUE), distribution type and
transform information.}
}}

