% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rca}
\alias{rca}
\title{Generate random number from a correlated accumulator model}
\usage{
rca(n, A, b, t0, mean_v, sd_v, st0 = 0, corr_v = 0,
  return_ttf = FALSE)
}
\arguments{
\item{n}{number of observatoins}

\item{A}{start point variability}

\item{b}{decision threshold}

\item{t0}{non-decision time}

\item{mean_v}{mean drift rate vector}

\item{sd_v}{standard deviation of the drift rates}

\item{st0}{non-decision time variability}

\item{corr_v}{correlations among accumulators}

\item{return_ttf}{a Boolean switch to return time to finish matrix}
}
\description{
Generate random number from a correlated accumulator model
}
