% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/likelihood.R
\name{remove_t0}
\alias{remove_t0}
\alias{removet0}
\title{Substract nondecision times from response times}
\usage{
remove_t0(x, t0)

removet0(x, t0)
}
\arguments{
\item{x}{a response time vector}

\item{t0}{a numeric scalar}
}
\value{
a vector
}
\description{
This function minus t0's from RTs.
}
\examples{
rt <- rlnorm(10) + .2
dt <- removet0(rt, .2)
all.equal(dt, rt - .2)
}
