% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_dgmc}
\alias{run_dgmc}
\title{Fit a Bayesian Model to Data}
\usage{
run_dgmc(samples, force, report, pm, qm, gammamult, ncore, ngroup)
}
\arguments{
\item{samples}{a initialized sample}

\item{force}{PDA re-calculate interval}

\item{report}{progress report intervel}

\item{pm}{probability of migration}

\item{qm}{probability of mutation}

\item{ncore}{number of CPU cores}

\item{ngroup}{number of independent groups}
}
\value{
Bayesian samples
}
\description{
run_dgmc runs Bayesian modelling with DGMC sampler.
}
