% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{run_many}
\alias{run_many}
\title{Fit a Bayesian Model to multiple Participants}
\usage{
run_many(samples, report, ncore, pm, pm0, qm, gammamult, ngroup, force,
  sampler, slice)
}
\arguments{
\item{samples}{a initialized samples list. Each element should contain
samples for a participant.}

\item{report}{progress report intervel}

\item{ncore}{number of CPU cores}

\item{pm}{probability of migration}

\item{pm0}{probability of migration}

\item{qm}{probability of mutation}

\item{gammamult}{turning parameter for crossover sampler}

\item{ngroup}{number of independent groups}

\item{force}{PDA re-calculate interval}

\item{sampler}{a string indicating to use which sampler}

\item{slice}{a Boolean switch to do slice sampling}
}
\description{
Use either DE-MCMC or DGMC sampler to fit independent Bayesian model to
many participants.
}
