% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{run_one}
\alias{run_one}
\title{Fit a Bayesian Model to a Single Participant}
\usage{
run_one(samples, report, pm, pm0, qm, gammamult, ngroup, force, sampler,
  slice)
}
\arguments{
\item{samples}{a initialized sample}

\item{report}{progress report intervel}

\item{pm}{probability of migration type 1}

\item{pm0}{probability of migration type 2 (Turner et al., 2013)}

\item{qm}{probability of mutation}

\item{gammamult}{tuning parameter of the crossover sampler}

\item{ngroup}{number of independent groups}

\item{force}{PDA re-calculate interval}

\item{sampler}{a string indicating to use which sampler}

\item{slice}{a Boolean switch to debug blocked sampling}
}
\value{
posterior samples
}
\description{
Use either DE-MCMC or DGMC sampler to fit Bayesian model to a participant
}
