% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rvonmises}
\alias{rvonmises}
\alias{dvonmises}
\alias{pvonmises}
\title{Generate random deviates from a von Mises distribution}
\usage{
rvonmises(n, mu, kappa)

dvonmises(x, mu, kappa)

pvonmises(q, mu, kappa, tol = 1e-20)
}
\arguments{
\item{n}{number of observations. Must be a scalar.}

\item{mu}{mean direction of the distribution. Must be a scalar.}

\item{kappa}{concentration parameter. A positive value
for the concentration parameter of the distribution. Must be a scalar.}
}
\value{
a column vector
}
\description{
This function generates random numbers in radian unit from a von Mises
distribution using the location (ie mean) parameter, mu and the
concentration (ie precision) parameter kappa.
}
\details{
A random number for a circular normal distribution has the form:\cr
\deqn{f(theta; mu, kappa) = 1 / (2*pi*I0(kappa)) * exp(kappa*cos(theta-mu))}
theta is between 0 and 2*pi.

\code{I0(kappa)} in the normalizing constant is the modified Bessel
function of the first kind and order zero.
}
\examples{
n  <- 1e2
mu <- 0
k  <- 10

\dontrun{
vm1 <- circular:::RvonmisesRad(n, mu, k)
vm2 <- rvm(n, mu, k)
vm3 <- circular:::conversion.circular(circular:::circular(vm1))
vm4 <- circular:::conversion.circular(circular:::circular(vm2))
plot(vm3)
plot(vm4)
}
}
\references{
\enumerate{
Ulric Lund, Claudio Agostinelli, et al's  (2017). R package 'circular':
Circular Statistics (version 0.4-91).
\url{https://r-forge.r-project.org/projects/circular/}
}
}
