% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{score}
\alias{score}
\title{Scoring RT data}
\usage{
score(x, digits = 2)
}
\arguments{
\item{x}{a parameter vector}

\item{digits}{printing digits}
}
\description{
A convenient function to calculate mean, interquantile range, standard
deviation for correct and error RTs
}
