% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi.R
\name{set_up_new_samples}
\alias{set_up_new_samples}
\title{Set Theta Storage for Model}
\usage{
set_up_new_samples(theta_input)
}
\arguments{
\item{theta_input}{An S4 object containing parameters specifying
the dimension and other information for the storage.}
}
\value{
An S4 object of class `posterior` containing:
\itemize{
  \item The configured theta parameters
  \item Additional metadata needed for posterior calculations
}
}
\description{
Creates and configures a theta container for use in model fitting.
}
\details{
This function takes the configuration parameter needed to create a posterior
}
\examples{
\dontrun{
result <- set_up_new_samples(theta_input)
}
}
