% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spdf}
\alias{spdf}
\title{Simulated likelihood functions for RT models}
\usage{
spdf(x, RT, n, h_in, debug = FALSE)
}
\arguments{
\item{x}{empirical data}

\item{RT}{simulated resposne times}

\item{n}{number of model simulations}

\item{h_in}{kernel bandwidth}

\item{debug}{debugging?}
}
\value{
a vector
}
\description{
CPU-based PDA
}
