% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,posterior-method}
\alias{summary,list-method}
\alias{summary,hyper-method}
\title{ggdmc Summary Methods}
\usage{
summary(object, ...)

\S4method{summary}{posterior}(object, start = 1, end = NA,
  prob = c(0.025, 0.25, 0.5, 0.75, 0.975), recovery = FALSE, ps = NA,
  verbose = FALSE, digits = max(3, getOption("digits") - 3))

\S4method{summary}{list}(object, start = 1, end = NA, prob = c(0.025,
  0.25, 0.5, 0.75, 0.975), recovery = FALSE, ps = NA,
  verbose = FALSE, digits = max(3, getOption("digits") - 3))

\S4method{summary}{hyper}(object, hyper = TRUE, start = 1, end = NA,
  prob = c(0.025, 0.25, 0.5, 0.75, 0.975), recovery = FALSE, ps = NA,
  type = 1, verbose = FALSE, digits = max(3, getOption("digits") -
  3))
}
\arguments{
\item{object}{an object storing posterior samples.}

\item{...}{Additional argument passing via dot dot dot.}

\item{start}{start from which iteration.}

\item{end}{end at which iteration. For example, set
\code{start = 101} and \code{end = 1000}, instructs the function to
calculate from 101st to 1000th iteration.}

\item{prob}{a numeric vector, indicating the quantiles to calculate}

\item{recovery}{a Boolean switch indicating if samples are from a recovery
study.}

\item{ps}{true parameter values.  This is only for recovery studies}

\item{verbose}{print more information}

\item{digits}{printing digits}

\item{hyper}{a Boolean switch to plot hyper parameters}

\item{type}{calculate type 1 = location or type 2 = scale hyper parameters}
}
\description{
Summarise posterior samples. Note when recovery = TRUE, the prob vector
will be fixed at the default values.
}
\examples{
\dontrun{
est1 <- summary(hsam[[1]], FALSE)
est2 <- summary(hsam[[1]], FALSE, 1, 100)

est3 <- summary(hsam)
est4 <- summary(hsam, verbose = TRUE)
est5 <- summary(hsam, verbose = FALSE)

hest1 <- summary(hsam, TRUE)
}
}
