% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{summary_many}
\alias{summary_many}
\title{Summarise Multiple Posterior Objects}
\usage{
summary_many(objects, start = 1L, end = NULL, verbose = FALSE)
}
\arguments{
\item{objects}{A list of posterior objects to summarise}

\item{start}{First iteration to include in summary}

\item{end}{Last iteration to include in summary (NULL for all)}

\item{verbose}{Logical Whether to print summarised results. Default: FALSE}
}
\value{
Either a list of summary objects (if verbose=FALSE) or a matrix of 
means
}
\description{
Summarise Multiple Posterior Objects
}
\examples{
\dontrun{
# Assuming you have set up DMI, priors etc.
# First make sure that the subject level has converged
fits0 <- StartSampling(pop_dmis, pop_priors,
    sub_migration_prob = 0.05,
    thin = 8L, pop_debug = F, seed = 9032
)

# Then turn on the migration sampler at the population level
fits1 <- RestartSampling(fits0,
    pop_migration_prob = 0.02,
    sub_migration_prob = 0.00,
    thin = 4L, seed = 9032
)

# Turn down the migration sampler, so that we may not fall
# into local minimal
fits2 <- RestartSampling(fits1,
    pop_migration_prob = 0.01,
    sub_migration_prob = 0.00,
    thin = 2L, seed = 9032
)

fits <- fits2
phi <- RebuildHyper(fits)
thetas <- RebuildPosteriors(fits)
result <- summary_many(thetas)

# Return all details for every participant
result <- summary_many(thetas, verbose = TRUE)
}
}
