% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summary_mcmc_list}
\alias{summary_mcmc_list}
\title{Summary statistic for posterior samples}
\usage{
summary_mcmc_list(object, prob = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{posterior samples}

\item{prob}{summary quantile summary}

\item{...}{other arguments passing in}
}
\description{
Calculate summary statistics for posterior samples
}
