% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{theta.as.mcmc.list}
\alias{theta.as.mcmc.list}
\alias{phi.as.mcmc.list}
\title{Convert Theta to a mcmc List}
\usage{
theta.as.mcmc.list(samples, start = 1, end = NA, split = FALSE)

phi.as.mcmc.list(hyper, start = 1, end = NA, split = FALSE, thin = 1)
}
\arguments{
\item{samples}{a sample list}

\item{start}{start iteration}

\item{end}{end iteraton}

\item{split}{whether to divide one MCMC sequence into two sequences.}
}
\description{
\code{theta.as.mcmc.list} extracts data-level parameter array, \code{theta},
from a DMC sample and convert it to a \pkg{coda} mcmc.list.
\code{phi.as.mcmc.list} extracts hyper-level parameter array, \code{phi}
from a DMC sample and convert it to a \pkg{coda} mcmc.list.
When \code{split} switch is TRUE, the function doubles number of chains,
first and second half.
}
