% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{theta_input-class}
\alias{theta_input-class}
\title{An S4 class to store MCMC sampling parameters.}
\description{
This class holds parameters that control the Markov chain Monte Carlo (MCMC)
sampling process.
}
\details{
This class is used to encapsulate the key parameters that govern the
MCMC sampling process. These parameters are essential for controlling
the length of the MCMC run, how often to keep samples, and the
storage and reporting of the sampling process.
}
\section{Slots}{

\describe{
\item{\code{nmc}}{A integer specifying the total number of MCMC iterations
(including burn-in).}

\item{\code{nchain}}{A integer indicating the number of MCMC chains to run.}

\item{\code{thin}}{A integer indicating the thinning interval.  This
determines how often samples are kept (e.g., a `thin` of 10 keeps every
10th sample). Thinning can help reduce autocorrelation in the MCMC samples.}

\item{\code{nparameter}}{A integer storing the number of free parameter.}

\item{\code{pnames}}{A string vector storing the name of the free parameter.}

\item{\code{report_length}}{A integer specifying the interval at which
progress reports are printed during the MCMC run. When running parallel cores
on Windows, the print function may not work.}

\item{\code{max_init_attempts}}{A integer specifying the maximum number of
attempts to initialise an MC sampler via the \code{rprior()} random number
generation.}

\item{\code{is_print}}{A logical value specifying whether to print out the progress.
upon previous samples.}
}}

\section{Objects from the Class}{

An \code{theta_input} instance can be created using
the `setThetaInput` or the conventional `new("theta_input", ...)`
constructor, where `...` are named arguments corresponding to
the slots.
}

\examples{
if (requireNamespace("ggdmcModel", quietly = TRUE))
{
    BuildModel <- getFromNamespace("BuildModel", "ggdmcModel")
    model <- BuildModel(
        p_map = list(A = "1", B = "1", t0 = "1", mean_v = "M", sd_v = "1", st0 = "1"),
        match_map = list(M = list(s1 = "r1", s2 = "r2")),
        factors = list(S = c("s1", "s2")),
        constants = c(st0 = 0, sd_v = 1),
        accumulators = c("r1", "r2"),
        type = "lba"
    )
}

nmc <- 200
theta_input <- setThetaInput(nmc = nmc, pnames = model@pnames)
}
