% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmc_eam.R
\name{transform.ddm}
\alias{transform.ddm}
\alias{transform.lba_B}
\title{Transform Parameter Data Frame}
\usage{
\method{transform}{ddm}(`_data`, type.par.names = NULL, ...)

\method{transform}{lba_B}(`_data`, type.par.names = NULL, ...)
}
\arguments{
\item{_data}{original called \code{par.df}, which is a data frame.}

\item{type.par.names}{default is NULL}

\item{...}{other arguments}
}
\description{
\code{transfrom.ddm} transforms parameters to a form suitbale for the model
being used. Called inside of get.par.mat. "par.df" is a data frame of
parameters types , some of which may need to be transformed, or new
columns created, so that the full set of internal parameter types, specified
in "type.par.names", required by the type of evidence accumulation model
being used ("norm" etc.) is present.
}
\details{
\code{transform.lba_B} template setup for n-choice LBA,
B=b-A parameterization
\itemize{
  \item External parameters types: A, B, t0, mean_v, sd_v, st0 = 0 (optional)
  \item Internal parameters types: A, b, t0, mean_v, sd_v, st0 = 0 (optional)
}
}
\keyword{transform}

