% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trial_loglik_hier}
\alias{trial_loglik_hier}
\title{Extract trial log likelihoods}
\usage{
trial_loglik_hier(samples, thin = 1, verbose = FALSE)
}
\arguments{
\item{samples}{posterior samples}

\item{thin}{thinnng length}

\item{verbose}{whether print information}
}
\description{
This function simply run trial_loglik to loop through one subject after
another to extracts trial_log_likes from a list of subject fits and
concatanates the result into an array.
}
