% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmc_utils.R
\name{view}
\alias{view}
\title{view function}
\usage{
view(p.prior)
}
\arguments{
\item{p.prior}{a prior list}
}
\value{
a p.prior data frame
}
\description{
a convenient function to rearrange p.prior or an element in pp.prior as a
data frame to easily inspect. That is another function called View
(uppercase) in utils packages.
}
\examples{
# Use a single value to replace all missing values
pop.mean  <- c(a=1,  v.f1=1,  v.f2=.2, z=0.5, sz=0.3,  sv.f1=0.25,
    sv.f2=0.23, t0=0.3)
pop.scale <-c(a=0.2,v.f1=.2, v.f2=.2, z=0.1, sz=0.05, sv.f1=.05,
    sv.f2=.05,  t0=0.05)
p.prior <- prior.p.dmc(
  dists = rep("tnorm", 8),
  p1=pop.mean,
  p2=pop.scale,
  lower=c(0,-5, -5, 0, 0, 0, 0,0),
  upper=c(2, 5, 5, 1, 2, 2, 1, 1))
view(p.prior)
}

