% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ac}
\alias{ac}
\title{Calculate the autocorrelation of a vector}
\usage{
ac(x, nLags = 50)
}
\arguments{
\item{x}{a vector storing parameter values}

\item{nLags}{the maximum number of lags}
}
\value{
A data.frame
}
\description{
Calculate the autocorrelation of a vector.
}
\details{
Internal function used by \code{\link{ggs_autocorrelation}}.
}
\examples{
## Calculate the autocorrelation of a simple vector
}
