% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ac}
\alias{ac}
\title{Calculate autocorrelation}
\usage{
ac(x, nlag)
}
\arguments{
\item{x}{likelihood vector}

\item{nlag}{number of autocorrelation lags}
}
\description{
This function calculate autocorrealtion for a Markov Chain
}
