% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmc_plotting.R
\name{acf.dmc}
\alias{acf.dmc}
\title{Plot an Autocorrelation Matrix}
\usage{
acf.dmc(x, lag.max = NULL, start = 1)
}
\arguments{
\item{x}{a DMC sample/object}

\item{lag.max}{to make it compartible to \code{acf} of \code{stats} pacakge.
Default is NULL}

\item{start}{plot from which iteration. Default is from the first iteration.}
}
\description{
This is a wrapper function to plot a DMC object, using \pkg{ggmcmc}
\code{ggs_autocorrelation}
}

