% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autocor}
\alias{autocor}
\title{Autocorrelation Plot}
\usage{
autocor(x, start = 1, end = NA)
}
\arguments{
\item{x}{posterior samples}

\item{start}{start from which iteration.}

\item{end}{end at which iteration}
}
\description{
Plot the autocorrelation of posterior samples,
}
