% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autocorr}
\alias{autocorr}
\title{Autocorrelation Plot}
\usage{
autocorr(x, start = 1, end = NA, nLags = 50, subchain = FALSE,
  nsubchain = 3, chains = NA, pll = FALSE, ll = FALSE, ...)
}
\arguments{
\item{x}{posterior samples}

\item{start}{start from which iteration.}

\item{end}{end at which iteration}

\item{nLags}{the maximum number of lags.}

\item{subchain}{boolean switch convert only a subset of chains.}

\item{nsubchain}{the number of chains in the subset.}

\item{chains}{numerical indexs indicating which chains to plot.}

\item{pll}{a boolean switch for plotting parameter values or posterior log_likelihoods}

\item{ll}{a boolean switch for plotting parameter values or log_likelihoods}
}
\description{
Plot the autocorrelation of posterior samples,
}
