% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{BuildDMI}
\alias{BuildDMI}
\title{Bind data and models}
\usage{
BuildDMI(x, model, npda = 16384, bw = 0.01, gpuid = 0,
  debug = FALSE)
}
\arguments{
\item{x}{data in a data frame format}

\item{model}{a model object}

\item{npda}{number of model simulations}

\item{bw}{kernel bandwidth}

\item{gpuid}{GPU ID, indicating using which GPU cards on a machine with
multieple GPUs.}

\item{debug}{debugging switch}
}
\description{
Binding a data with a model object. This function also checks
if they are compatible and adds a \code{cell.index} and many other
attributes to the data frame.
}
