% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{BuildDMI}
\alias{BuildDMI}
\title{Bind data and models}
\usage{
BuildDMI(x, model)
}
\arguments{
\item{x}{data as in data frame}

\item{model}{a model object}
}
\value{
a data model instance
}
\description{
Binding a data set with a model object. The function also checks whether
they are compatible and adds attributes on a data model instance.
}
