% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc_model.R
\name{check_pvec}
\alias{check_pvec}
\title{Is a p.vector compatible with model?}
\usage{
check_pvec(pvec, model)
}
\arguments{
\item{pvec}{p.vector. I use pvec to avoid naming confound with S3 method.}

\item{model}{a model object}
}
\description{
Check if the user supplies a p.vector, incompatiable with the p.vector
created by \code{model.dmc}.
}
