% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{check_pvec}
\alias{check_pvec}
\title{Does a model object specify a correct p.vector}
\usage{
check_pvec(p.vector, model)
}
\arguments{
\item{p.vector}{parameter vector}

\item{model}{a model object}
}
\description{
Check a parameter vector
}
