% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{dbeta_lu}
\alias{dbeta_lu}
\title{A modified dbeta function}
\usage{
dbeta_lu(x, p1, p2, lower, upper, lg = FALSE)
}
\arguments{
\item{x}{quantile}

\item{p1}{shape1 parameter}

\item{p2}{shape2 parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{lg}{logical; if TRUE, return log density.}
}
\description{
A modified dbeta function
}
