% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{dcauchy_l}
\alias{dcauchy_l}
\alias{rcauchy_l}
\title{A slightly modified dcauchy and rcauchy functions}
\usage{
dcauchy_l(x, location, scale, log = FALSE)

rcauchy_l(n, location, scale)
}
\arguments{
\item{x}{posterior samples}

\item{location}{location parameter}

\item{scale}{scale parameter}

\item{log}{log density?}

\item{n}{number of simulations}
}
\description{
A slightly modified dcauchy and rcauchy functions
}
