% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcircle}
\alias{dcircle}
\alias{dcircle300}
\alias{rcircle}
\alias{rcircle_process}
\alias{r1d}
\title{The 2-dimension Diffusion Model}
\usage{
dcircle(RT, A, P, tmax, kmax, sz, nw)

dcircle300(P, tmax, kmax, sz, nw)

rcircle(n, P, tmax, h, nw)

rcircle_process(P, tmax, h)

r1d(P, tmax, h)
}
\arguments{
\item{P}{is a parameter vector, c(v1, v2, a, t0, sigma1, sigma2, eta1, eta2).
The sequence is important. v1 is the x-axis mean drift rate. v2 is the
y-axis mean drift rate. sigma1 is the x-axis within-trial drift rate SD.
sigma2 is the y-axix within-trial drift rate SD. a is decision threshold.
sigma1 and sigma2 must be 1 and identical, because this is what has been
thoroughly tested so far. Other values may return unknown results. t0
non-decision time.}

\item{tmax}{maximum time of the model}

\item{kmax}{the tuning parameter for Bessel function. Mostly 50.}

\item{nw}{the number of theta steps (w = 2 * pi / nw)}

\item{h, }{sz the number of time steps (h = tmax / sz). h is time step.
Mostly .1 ms.}

\item{x, }{n x is a data.matrix with first column of RT and a second column
of R. n is the numbers of observation.}
}
\value{
rcircle returns a n x 2 matrix. Each row is an [RT R] trial.
dcircle returns a n vector.
}
\description{
Density, random generation for the 2-D diffusion model.
}
\details{
The model has the main parameters, v1, v2, eta1, eta2, a, sigma, and t0.
tmax, kmax, sz and nw are tuning parameters for determining the set.
dcircle300 produces PDF table and others.
}
\examples{
## TODO examples
}
