% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{dconstant}
\alias{dconstant}
\title{A pseudo constant function to get constant densities}
\usage{
dconstant(x, p1, p2, lower, upper, lg = FALSE)
}
\arguments{
\item{x}{quantile}

\item{p1}{constant value}

\item{p2}{unused argument}

\item{lower}{dummy varlable}

\item{upper}{dummy varlable}

\item{lg}{log density?}
}
\description{
Used with constant prior
}
