% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{deviance.model}
\alias{deviance.model}
\title{Calculate the statistics of model complexity}
\usage{
\method{deviance}{model}(object, ...)
}
\arguments{
\item{object}{posterior samples}

\item{...}{other plotting arguments passing through dot dot dot.}
}
\description{
Calculate deviance for a model object for which a
log-likelihood value can be obtained, according to the formula
-2*log-likelihood.
}
