% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{dgamma_l}
\alias{dgamma_l}
\title{A modified dgamma function}
\usage{
dgamma_l(x, p1, p2, lower, upper, lg = FALSE)
}
\arguments{
\item{x}{quantile}

\item{p1}{shape parameter}

\item{p2}{scale parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{lg}{log density?}
}
\description{
A modified dgamma function
}
