% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{DIC}
\alias{DIC}
\alias{BPIC}
\title{Deviance information criteria}
\usage{
DIC(object, ...)

BPIC(object, ...)
}
\arguments{
\item{object}{posterior samples}

\item{...}{other plotting arguments passing through dot dot dot.}
}
\description{
Calculate DIC and BPIC.
}
