% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{dlnorm_l}
\alias{dlnorm_l}
\title{A modified dlnorm functions}
\usage{
dlnorm_l(x, p1, p2, lower, upper, lg = FALSE)
}
\arguments{
\item{x}{quantile}

\item{p1}{meanlog parameter}

\item{p2}{sdlog parameter}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{lg}{log density?}
}
\description{
A modified dlnorm functions
}
