% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{effectiveSize}
\alias{effectiveSize}
\title{Effective Sample Size for Estimating the Mean}
\usage{
effectiveSize(x, hyper = FALSE, start = 1, end = NA, digits = 0,
  verbose = FALSE)
}
\arguments{
\item{x}{a samples object}

\item{hyper}{a switch to extract hyper attribute and calculate it}

\item{start}{starting iteration}

\item{end}{ending iteraton}

\item{digits}{printing digits}
}
\description{
\code{effectiveSize} calls \pkg{coda} effectiveSize to effective size
for either single or multiple subjects. It can calculate at the data or
hyper level, too.
}
\examples{
#################################40
## effectiveSize example
#################################40
}
