% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Retrieve information of operating system}
\usage{
get_os()
}
\description{
A wrapper function to extract system information from \code{Sys.info}
and \code{.Platform}
}
\examples{
get_os()
## sysname
## "linux"
}
