% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetPNames}
\alias{GetPNames}
\title{Extract parameter names from a model object}
\usage{
GetPNames(model)
}
\arguments{
\item{model}{a model object}
}
\description{
Extract parameter names from a model object
}
