% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{package}
\name{ggdmc}
\alias{ggdmc}
\alias{ggdmc-package}
\title{Bayeisan computation of response time models}
\description{
\pkg{ggdmc} uses the population-based Markov chain Monte Carlo to
conduct Bayesian computation on cognitive models.
}
\references{
Heathcote, A., Lin, Y.-S., Reynolds, A., Strickland, L., Gretton, M. &
Matzke, D., (2018). Dynamic model of choice.
\emph{Behavior Research Methods}.
https://doi.org/10.3758/s13428-018-1067-y. \cr

Turner, B. M., & Sederberg P. B. (2012). Approximate Bayesian computation
with differential evolution, \emph{Journal of Mathematical Psychology}, 56,
375--385. \cr

Ter Braak (2006). A Markov Chain Monte Carlo version of the genetic
algorithm Differential Evolution: easy Bayesian computing for real
parameter spaces. \emph{Statistics and Computing}, 16, 239-249.
}
\author{
Yi-Shin Lin <yishinlin001@gmail.com> \cr
Andrew Heathcote <andrew.heathcote@utas.edu.au>
}
\keyword{package}
