% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{iseffective}
\alias{iseffective}
\title{Model checking functions}
\usage{
iseffective(x, minN, nfun, verbose = FALSE)
}
\arguments{
\item{x}{posterior samples}

\item{minN}{specify the size of minimal effective samples}

\item{nfun}{specify to use the \code{mean} or \code{median} function to
calculate effective samples}

\item{verbose}{print more information}
}
\description{
The function tests whether we have drawn enough samples.
}
