% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{isstuck}
\alias{isstuck}
\title{Four pMCMC checking tools for automatic sampling}
\usage{
isstuck(x, cut = 10, verbose = FALSE)
}
\arguments{
\item{x}{posterior samples}

\item{cut}{the criteria for suggesting abnormal chains found}

\item{verbose}{print more information}
}
\description{
The function tests whether MCMC chains encounter a parameter region
difficult to search (ie get stuck):
}
