% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{model-class}
\alias{model-class}
\title{An S4 class of the process model.}
\description{
The class is to represent a process model, e.g., a DDM, a LBA model, a PM
model, or a CDDM.
}
\section{Slots}{

\describe{
\item{\code{model}}{A 3-D model array. Dimension one stores the combinations of the
factor levels and response types (when discrete), dimension two stores
parameters, and dimension three stores response types.}

\item{\code{all.par}}{all parameters}

\item{\code{p.vector}}{parameter vector, excluding constant parameters}

\item{\code{par.names}}{parameter names / labels}

\item{\code{type}}{model type}

\item{\code{factors}}{a list of factors and their levels}

\item{\code{responses}}{response types}

\item{\code{constants}}{constant parameters}

\item{\code{posdrift}}{a Boolean switch indicating whether drift rates must be
positive}

\item{\code{n1.order}}{node 1 ordering. This is only for the LBA model}

\item{\code{match.cell}}{an indicator matrix storing whether a particular trial
matches a cell}

\item{\code{match.map}}{a mapping mechanism for calculating whether a trial matches
a positive boundary / accumulator or a negative boundary / accumulator.}

\item{\code{dimnames}}{dimension names of the model array}

\item{\code{pnames}}{parameter names}

\item{\code{npar}}{number of parameters}
}}

