% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{methods}
\name{names,prior-method}
\alias{names,prior-method}
\title{The Parameter Names of a Prior Object}
\usage{
\S4method{names}{prior}(x)
}
\arguments{
\item{x}{a prior object.}
}
\value{
a string vector
}
\description{
Extract parameter names from a prior object.
}
