% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{posterior-class}
\alias{posterior-class}
\title{An S4 class to represent an object storing posterior samples at the
participant level. Posterior samples storing both the participant and the
hyper lever are represented by an S4 class hyper}
\description{
An S4 class to represent an object storing posterior samples at the
participant level. Posterior samples storing both the participant and the
hyper lever are represented by an S4 class hyper
}
\section{Slots}{

\describe{
\item{\code{theta}}{posterior samples for one-participant fit.}

\item{\code{summed_log_prior}}{summed log prior likelihoods.}

\item{\code{log_likelihoods}}{log likelihoods}

\item{\code{dmi}}{a S4 object of data model instance}

\item{\code{prior}}{a S4 prior object}

\item{\code{start}}{the index of starting sample}

\item{\code{npar}}{number of parameters}

\item{\code{pnames}}{parameter names}

\item{\code{nmc}}{number of Monte Carlo samples}

\item{\code{thin}}{thinning length}

\item{\code{nchain}}{number of Markov chains}
}}

\seealso{
\code{\link{hyper-class}}
}
