% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random}
\alias{random}
\title{Generate random numbers}
\usage{
random(type, pmat, n, seed = NULL)
}
\arguments{
\item{type}{a character string indicating the model type}

\item{pmat}{a matrix of response x parameter}

\item{n}{number of simulations}

\item{seed}{an integer specifying the random seed}
}
\description{
A wrapper function for the model type, \code{rd}, \code{norm}.
}
