% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random}
\alias{random}
\title{Generate random numbers}
\usage{
random(type, pmat, n, seed = NULL)
}
\arguments{
\item{type}{a character string of the model type}

\item{pmat}{a matrix of response x parameter}

\item{n}{number of observations}

\item{seed}{an integer specifying a random seed}
}
\description{
A wrapper function for generating random numbers of either
the model type, \code{rd}, or \code{norm}.
}
