% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/ggdmc_likelihood.R
\name{remove_t0}
\alias{remove_t0}
\alias{removet0}
\title{Remove t0}
\usage{
remove_t0(rt, t0)

removet0(RT, t0)
}
\arguments{
\item{t0}{a numeric scalar}

\item{RT, rt}{a numeric vector}
}
\value{
a vector
}
\description{
This function minus t0 from RT.
}
\examples{
rt <- rlnorm(10) + .2
dt <- ggdmc::removet0(rt, .2)
all.equal(dt, rt - .2)
}
