% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rlba_norm}
\alias{rlba_norm}
\title{Generate Random Deviates of the LBA Distribution}
\usage{
rlba_norm(n, A, b, mean_v, sd_v, t0, st0, posdrift)
}
\arguments{
\item{n}{is the numbers of observation.}

\item{A}{start point upper bound, a vector of a scalar.}

\item{b}{decision threshold, a vector or a scalar.}

\item{mean_v}{mean drift rate vector}

\item{sd_v}{standard deviation of drift rate vector}

\item{t0}{nondecision time, a vector.}

\item{st0}{nondecision time variation, a vector.}

\item{posdrift}{if exclude negative drift rates}
}
\value{
a n x 2 matrix of RTs (first column) and responses (second column).
}
\description{
\code{rlba_norm}, only slightly faster than \code{maker}, calls C++
function directly.
}
