% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmc_sampling.R
\name{run}
\alias{run}
\alias{CheckConverged}
\alias{StuckTests}
\alias{FlatTests}
\alias{MixTests}
\alias{LengthTests}
\alias{GetPLL}
\title{Fit a fixed-effect or hierarchical Bayesian models}
\usage{
run(samples, report = 100, ncore = 1, pm = 0, qm = 0,
  gamma.mult = 2.38, ngroup = 6, force = FALSE, sampler = "DE-MCMC",
  debug = FALSE)

CheckConverged(samples)

StuckTests(samples, verbose = FALSE, cut = 10)

FlatTests(samples, p1 = 1/3, p2 = 1/3, cut.location = 0.25,
  cut.scale = Inf, verbose = FALSE)

MixTests(samples, verbose = FALSE, cut = 1.01, split = TRUE)

LengthTests(samples, minN, nfun, verbose = FALSE)

GetPLL(fit, pars, start, end)
}
\arguments{
\item{samples}{a sample list generated by calling DMC's samples.dmc.}

\item{report}{how many iterations to return a report}

\item{gamma.mult}{a DEMC tuning parameter, affecting the size of jump}

\item{force}{Set force to FALSE for turning off force resampling. Set it
as an integer 1 to 10, forcing to resample a new parameter proposal every,
e.g., 1, 2, 3 step.}

\item{cores}{a switch for computing the prob density for each trial in
parallel. Turn it on by setting any number > 1.}

\item{p.migrate}{set it greater than 0 to use migration samplers. For example
p.migrate=0.05 will use migration in 5\% chance.}

\item{farjump}{No funciton for compatibility reason}
}
\value{
a DMC sample with class c("list", "dmc")
}
\description{
This function fit a hierarchical or a fixed-effect model, using Bayeisan
sampling.  We use pMCMC, with a suite of DE-MCMC, DGMC, and simply,
crossover (i.e., DE-MC), mutation, or migration operators. Note that
the latter two operators essentially are random-walk Metroplolis, so they
will be very inefficient, if been applied alone, even with our fast C++
implementation.
}
