% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmc_sampling.R
\name{StartNewsamples}
\alias{StartNewsamples}
\title{Initialize New Samples}
\usage{
StartNewsamples(nmc, data = NULL, p.prior = NULL, thin = 1,
  theta1 = NULL, rp = 0.001, nchain = NULL)
}
\arguments{
\item{nmc}{numbers of MCMC iterations/samples}

\item{data}{a data-model instance from \code{BindDataModel}.}

\item{p.prior}{parameter priors}

\item{thin}{thinning length.}

\item{theta1}{user-supplied theta}

\item{rp}{DE-MCMC tuning parameter. Default is 0.001.}

\item{nchain}{numbers of Markov chains. Default is 3 times numbers of
parameter.}

\item{samples}{a posterior sample. If data is not supplied, the function
will look for \code{samples}.}
}
\description{
Initialize New Samples
}
\examples{
m1 <- BuildModel(
    p.map     = list(a="1",v="F",z="1",d="1",sz="1",sv="1", t0="1",st0="1"),
    constants = c(st0=0,d=0),
    match.map = list(M=list(s1="r1",s2="r2")),
    factors   = list(S=c("s1","s2"), F=c("f1", "f2")),
    responses = c("r1","r2"),
    type      = "rd")

## m1 is "dmc" class
class(m1)
## [1] "dmc"

pVec <- c(a=1, v.f1=1, v.f2=1.5, z=0.5, sz=0.25, sv=0.2,t0=.15)
dat  <- simulate(m1, nsim=1e2, p.vector=pVec)
str(dat)
## 'data.frame':	400 obs. of  4 variables:
## $ S : Factor w/ 2 levels "s1","s2": 1 1 1 1 1 1 1 1 1 1 ...
## $ F : Factor w/ 2 levels "f1","f2": 1 1 1 1 1 1 1 1 1 1 ...
## $ R : Factor w/ 2 levels "r1","r2": 1 1 1 2 1 1 1 1 2 1 ...
## $ RT: num  0.26 0.255 0.572 0.25 0.518 ...

## mdi1 is dmc as well as data.frame class
mdi1 <- BindDataModel(dat, m1)

p.prior <- prior.p.dmc(
   dists = rep("tnorm", 7),
   p1    = c(a=2,  v.f1=2.5, v.f2=1.25, z=.5, sz=.3, sv=1,  t0=.3),
   p2    = c(a=.5, v.f1=.5,  v.f2=.35,  z=.1, sz=.1, sv=.3, t0=.05),
   lower = c(0,-5, -5, 0, 0, 0, 0),
   upper = c(5, 7,  7, 2, 2, 2, 2))

## Set up a new DMC sample with 200 iteration. The default thinning length
## is 1
samples0 <- samples(nmc=50, p.prior=p.prior, data=mdi1)
samples0$nmc
## [1] 50

## Run a fixed-effect model with 5\% chance of using migration sampler
## samples0 <- run(samples0, p.migrate=.05)

## Add 200 more iteration on to sample0
samples1 <- samples(nmc=50, p.prior=p.prior, samples=samples0, add=TRUE)
## samples1$nmc
## [1] 100
}
