% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{StartNewsamples}
\alias{StartNewsamples}
\alias{run}
\title{Start new model fits}
\usage{
StartNewsamples(data, prior = NULL, nmc = 500, thin = 1,
  nchain = NULL, report = 100, rp = 0.001, gammamult = 2.38,
  pm0 = 0.05, pm1 = 0.05, block = TRUE, ncore = 1)

run(samples, nmc = 500, thin = 1, report = 100, rp = 0.001,
  gammamult = 2.38, pm0 = 0, pm1 = 0, block = TRUE, ncore = 1,
  add = FALSE)
}
\arguments{
\item{data}{data model instance(s)}

\item{prior}{prior distributions. For hierarchical model, this must be a
list with three sets of prior distributions. Each is named, "pprior",
"location", and "scale".}

\item{nmc}{number of Monte Carlo sampling}

\item{thin}{thinning length}

\item{nchain}{number of chains}

\item{report}{progress report interval}

\item{rp}{tuning parameter 1}

\item{gammamult}{tuning parameter 2, step size of parameter search}

\item{pm0}{probability of migration type 0 (Hu & Tsui, 2010)}

\item{pm1}{probability of migration type 1 (Turner et al., 2013)}

\item{block}{Only for hierarchical modeling. Update one parameter at a time}

\item{ncore}{Only for non-hierarchical modeling with many subjects.}

\item{samples}{model fit(s).}

\item{add}{Boolean whether to add new samples}
}
\description{
Modelling a hierarchical or a fixed-effect model, using Bayeisan
optimisation.  We use a specific type of pMCMC algorithm, the DE-MCMC,
which includes crossover and two different migration operators. The latter
two operators essentially are population-version random-walk Metroplolis.
They will be less efficient to locate a target parameter space, if been
applied alone.
}
