% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summary_hyper}
\alias{summary_hyper}
\alias{summary_one}
\alias{summary_many}
\alias{summary_recoverone}
\alias{summary_recovermany}
\alias{summary_recoverhyper}
\alias{summary.model}
\title{Summarise a DMC Sample with One Participant}
\usage{
summary_hyper(object, start, end, hmeans, hci, prob, digits)

summary_one(object, start, end)

summary_many(object, start, end, verbose)

summary_recoverone(object, start, end, ps, digits, verbose)

summary_recovermany(object, start, end, ps, digits)

summary_recoverhyper(object, start, end, ps, type, digits, verbose)

\method{summary}{model}(object, hyper = FALSE, start = 1, end = NA,
  hmeans = FALSE, hci = FALSE, prob = c("2.5\%", "50\%", "97.5\%"),
  recovery = FALSE, ps = NA, type = 1, verbose = FALSE, ...,
  digits = 2)
}
\arguments{
\item{object}{a model samples}

\item{start}{summarise from which MCMC iteration. Default uses the first
iteration.}

\item{end}{summarise to the end of MCMC iteration. For example, set
\code{start=101} and \code{end=1000}, instructs the function to calculate
from 101 to 1000 iteration. Default uses the last iteration.}

\item{...}{other arguments}
}
\description{
Call coda package to summarise the model parameters in a DMC samples
}
\examples{
model <- BuildModel(p.map = list(A="1", B="R", t0="1",
mean_v=c("F", "M"), sd_v="M", st0="1"),
match.map = list(M = list(s1=1, s2=2)),
factors   = list(S = c("s1", "s2"),F = c("f1", "f2")),
constants = c(sd_v.false = 1, st0 = 0),
responses = c("r1", "r2"),
type      = "norm")
npar <- length(GetPNames(model))
}
