% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc_analysis.R
\name{theta.as.mcmc.list}
\alias{theta.as.mcmc.list}
\title{Convert Theta to a mcmc List}
\usage{
theta.as.mcmc.list(samples, start = 1, end = NA)
}
\arguments{
\item{samples}{a DMC sample}

\item{start}{start iteration}

\item{end}{end iteraton}
}
\description{
\code{theta.as.mcmc.list} extract theta from a DMC sample and convert it to
a \pkg{coda} mcmc.list object.
}

