% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{unstick_one}
\alias{unstick_one}
\title{Unstick posterios samples (One subject)}
\usage{
unstick_one(x, bad)
}
\arguments{
\item{x}{posterior samples}

\item{bad}{a numeric vector, indicating which chains to remove}
}
\description{
Unstick posterios samples (One subject)
}
